refactor: extract block computation logic into useComputedBlocks hook
Move the useEffect + useState block computation out of StackedBlockChart
into a dedicated useComputedBlocks hook so the component focuses solely
on rendering.
