+++ src/stacked-block-chart.tsx
@@ -1,20 +1,8 @@
-import { ComponentPropsWithRef, forwardRef, useEffect, useState } from "react";
+import { ComponentPropsWithRef, forwardRef } from "react";
 import Block from "./block";
 import BlockLabels from "./block-labels";
 import Legend from "./legend";
-import {
-  BlockDatum,
-  addXFluctuation,
-  adjustSameValueBlocks,
-  adjustTotalHeight,
-  alignToBottom,
-  calcPercentage,
-  calcWidthsAndHeights,
-  calcXPositions,
-  calcYPositions,
-  createInitialBlockDatum,
-  modifyOrderByType,
-} from "./compute-blocks";
+import { useComputedBlocks } from "./use-computed-blocks";
 import { useRandomGenerator } from "./use-random-generator";
 
 export type StackType = "stable-balanced" | "unstable-inverted" | "shuffled";
@@ -48,43 +36,12 @@ export const StackedBlockChart = forwardRef<
   ) => {
     const svgWidth = 400;
     const svgHeight = 300;
-    const blocksOffsetX = 40;
     const legendWidth = 100;
     const legendItemHeight = 16;
     const legendPaddingTop = 10;
     const legendPaddingRight = 10;
-    const [blocks, setBlocks] = useState<BlockDatum[]>([]);
-    const [legendItems, setLegendItems] = useState<
-      { name: string; color: string }[]
-    >([]);
     const { getRandomInt } = useRandomGenerator(seed);
-
-    useEffect(() => {
-      const initialBlocks = data.map(createInitialBlockDatum);
-      const total = initialBlocks.reduce((acc, d) => acc + d.value, 0);
-      const svgCenterX = (svgWidth - legendWidth) / 2 - blocksOffsetX;
-
-      const ops: ((b: BlockDatum[]) => BlockDatum[])[] = [
-        (b) => b.map((datum) => calcPercentage(datum, total)),
-        (b) => b.sort((a, b) => a.percentage - b.percentage),
-        (b) => calcWidthsAndHeights(b, getRandomInt, { multiple: 100 }),
-        (b) => adjustSameValueBlocks(b),
-        (b) => adjustTotalHeight(b, svgHeight),
-        (b) => modifyOrderByType(b, stackType, getRandomInt),
-        (b) => calcYPositions(b),
-        (b) => calcXPositions(b, svgCenterX),
-        (b) => addXFluctuation(b, getRandomInt),
-        (b) => alignToBottom(b, svgHeight),
-      ];
-
-      const blocks = ops.reduce((acc, op) => op(acc), initialBlocks);
-      const legendItems = blocks.map((d) => {
-        return { name: d.name, color: d.fill };
-      });
-
-      setBlocks(blocks);
-      setLegendItems(legendItems);
-    }, [stackType, data, getRandomInt]);
+    const { blocks, legendItems } = useComputedBlocks(data, stackType, getRandomInt);
 
     return (
       <>
